api/requirements.txt
selenium==4.27.1
google-cloud-storage==2.19.0
google-cloud-bigquery==3.29.0
PyYAML==6.0.2
webdriver-manager==4.0.2
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
fastparquet==2024.11.0
pandas==2.2.3
fsspec==2024.12.0
pyarrow==19.0.0
isort==5.13.2
black==24.10.0
ruff==0.9.2
pytest==8.3.4
